+++ addon/mixins/select-box/focusable.js
@@ -12,7 +12,7 @@ export default Mixin.create({
     if (isPresent(this.tabindex)) {
       set(this, 'tabIndex', this.tabindex);
     } else if (!isPresent(this.tabIndex)) {
-      set(this, 'tabIndex', 0);
+      set(this, 'tabIndex', '0');
     }
   },